fix: change from a global ack counter to a session one

+++ depChain/src/main/java/depchain/network/Message.java
@@ -14,10 +14,10 @@ public class Message implements Serializable {
     public final String value; // The candidate value (e.g., the string to append).
     public final int senderId; // The sender's ID (for clients, use a distinct range).
     public final byte[] signature; // Signature over the message content (computed by sender).
-    public long nonce; // nonce for the message (computed by sender).
+    public int nonce; // nonce for the message (computed by sender).
     public ByteArrayWrapper sessionKey; // session key for the message (computed by sender).
 
-    public Message(Type type, int epoch, String value, int senderId, byte[] signature, long nonce) {
+    public Message(Type type, int epoch, String value, int senderId, byte[] signature, int nonce) {
         this.type = type;
         this.epoch = epoch;
         this.value = value;
@@ -26,7 +26,7 @@ public class Message implements Serializable {
         this.nonce = nonce;
     }
 
-    public Message(Type type, int epoch, int senderId, byte[] signature, long nonce, ByteArrayWrapper sessionKey) {
+    public Message(Type type, int epoch, int senderId, byte[] signature, int nonce, ByteArrayWrapper sessionKey) {
         this.type = type;
         this.epoch = epoch;
         this.value = ""; // Initialize value with an empty string or any default value
@@ -36,7 +36,7 @@ public class Message implements Serializable {
         this.sessionKey = sessionKey;
     }
 
-    public void setNonce(long nonce) {
+    public void setNonce(int nonce) {
         this.nonce = nonce;
     }
